Show chores button now appears upon logging in new user
Fixes bug where if one user logs in, shows their chores, and then
logs out, when the next user logs in, they don't see the show chores button.

The show chores button appears but does not actually work right now.
It's not showing chores. Moving on to fixing that now.