Confirms that the right user and chore objects are being returned -- they're just not being displayed
Isolating the problem.
This means it's probably a combination of jQuery and HTML/CSS
that's causing the bug.

Right now, if a user signs in, shows chores, and signs out, and then
a second user signs in, the system is showing them the show-chore
button, and the API request is returning their chores.
But the chores are not displaying to the user.